Create gens & write modules
